codechef march 2

+++ lossprofit.java
@@ -7,17 +7,17 @@ class LossProfit{
   int t = sc.nextInt();
   for(int j=0;j<t;j++){
   int n = sc.nextInt();
-  float ans = 0f;
+  double ans = 0.0;
   for(int i=0;i<n;i++){
    int price = sc.nextInt();
    int quantity = sc.nextInt();
    int dis = sc.nextInt();
-   float mul = 0f;
-   float disn = 0f;
+   double mul = 0.0;
+   double disn = 0.0;
    if(dis > 0){
-     disn = dis/100f;
-     mul = disn*(float)price;
-     ans = (((float)price - (float)(price+mul)*(1-disn))*quantity)+ans;
+     disn = dis/100.0;
+     mul = disn*(double)price;
+     ans = (((double)price - (double)(price+mul)*(1-disn))*quantity)+ans;
    }
   }
  System.out.println(ans);
